# swagger.yaml
openapi: 3.0.0
paths:
...
  /api/access-control/roles:
    get:
      operationId: AccessControlController_findAllRoles
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: AccessControlController_createRole
      parameters: []
      responses:
        "201":
          description: ""
  /api/access-control/roles/{id}:
    get:
      operationId: AccessControlController_findOneRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    put:
      operationId: AccessControlController_updateRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    delete:
      operationId: AccessControlController_deleteRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
...
  /api/employees:
    post:
      operationId: EmployeeController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeDto"
      responses:
        "201":
          description: ""
    get:
      operationId: EmployeeController_findAll
      parameters: []
      responses:
        "200":
          description: ""
...
  /api/employees/{id}/manager-role:
    delete:
      operationId: EmployeeController_revokeManager
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/employees/stats/count:
    get:
      operationId: EmployeeController_getCount
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/employees/{id}:
    get:
      operationId: EmployeeController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    put:
      operationId: EmployeeController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeDto"
      responses:
        "200":
          description: ""
    delete:
      operationId: EmployeeController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
...
  /api/departments:
    post:
      operationId: DepartmentController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentDto"
      responses:
        "201":
          description: ""
    get:
      operationId: DepartmentController_findAll
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/departments/{id}:
    get:
      operationId: DepartmentController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    patch:
      operationId: DepartmentController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentDto"
      responses:
        "200":
          description: ""
    delete:
      operationId: DepartmentController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
...
  /api/finance/payroll:
    post:
      operationId: FinanceController_createPayroll
      parameters: []
      responses:
        "201":
          description: ""
    get:
      operationId: FinanceController_findAllPayroll
      parameters:
        - name: employeeId
          required: true
          in: query
          schema:
            type: string
        - name: companyId
          required: true
          in: query
          schema:
            type: string
        - name: month
          required: true
          in: query
          schema:
            type: number
        - name: year
          required: true
          in: query
          schema:
            type: number
      responses:
        "200":
          description: ""
...
  /api/finance/expenses:
    post:
      operationId: FinanceController_createExpense
      parameters: []
      responses:
        "201":
          description: ""
    get:
      operationId: FinanceController_findAllExpenses
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
        - name: currency
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/finance/expenses/{id}:
    delete:
      operationId: FinanceController_deleteExpense
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/finance/payroll/{id}/slip:
    get:
      operationId: FinanceController_getPayrollSlip
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
...
  /api/finance/summary:
    get:
      operationId: FinanceController_getSummary
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
        - name: month
          required: true
          in: query
          schema:
            type: number
        - name: year
          required: true
          in: query
          schema:
            type: number
        - name: currency
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/finance/payroll/{id}:
    patch:
      operationId: FinanceController_updatePayroll
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    delete:
      operationId: FinanceController_deletePayroll
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  /api/finance/payroll/{id}/send-email:
    post:
      operationId: FinanceController_sendPayrollEmail
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
...
  /api/email-templates/{id}:
    get:
      operationId: EmailTemplatesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    patch:
      operationId: EmailTemplatesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    delete:
      operationId: EmailTemplatesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
info:
  title: Difmo CRM API
  description: The Difmo CRM API description
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.difmocrm.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateEmployeeDto:
      type: object
      properties: {}
    UpdateEmployeeDto:
      type: object
      properties: {}
    CreateDepartmentDto:
      type: object
      properties: {}
    UpdateDepartmentDto:
      type: object
      properties: {}
...